# docs/openapi.yaml
openapi: 3.0.0
info:
  title: superglue AI API
  version: 1.0.0
  description: API for running superglue AI tools
  contact:
    name: superglue AI Support
    email: stefan@superglue.ai
servers:
  - url: https://api.superglue.ai/v1
    description: Production
security:
  - ApiKeyAuth: []
paths:
...
  /end-users:
    get:
      summary: List end users
      operationId: listEndUsers
      tags:
        - End Users (Enterprise)
      description: |
        **Enterprise feature.**
        
        List all end users in your organization. End users are external users who can connect their own credentials to multi-tenancy systems via the portal.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        "200":
          description: List of end users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/EndUser"
                  page:
                    type: integer
                  limit:
                    type: integer
                  total:
                    type: integer
                  hasMore:
                    type: boolean
    post:
      summary: Create or update end user
      operationId: upsertEndUser
      tags:
        - End Users (Enterprise)
      description: |
        **Enterprise feature.**
        
        Create a new end user or update an existing one (identified by externalId).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                externalId:
                  type: string
                  description: Your application's user ID (auto-generated if not provided)
                  example: user-123
                email:
                  type: string
                  format: email
                  example: user@example.com
                name:
                  type: string
                  example: John Doe
                allowedSystems:
                  type: array
                  description: Array of system IDs, or ["*"] for all systems
                  items:
                    type: string
                  example: ["*"]
      responses:
        "200":
          description: End user created or updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/EndUser"
  /end-users/{endUserId}:
    get:
      summary: Get end user details
      operationId: getEndUser
      tags:
        - End Users (Enterprise)
      description: |
        **Enterprise feature.**
        
        Get end user details including credential connection status for all systems.
      parameters:
        - name: endUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: End user details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: "#/components/schemas/EndUser"
                      - type: object
                        properties:
                          credentials:
                            type: array
                            description: Credential status for each system
                            items:
                              type: object
                              properties:
                                systemId:
                                  type: string
                                systemName:
                                  type: string
                                hasCredentials:
                                  type: boolean
        "404":
          description: End user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete end user
      operationId: deleteEndUser
      tags:
        - End Users (Enterprise)
      description: |
        **Enterprise feature.**
        
        Delete an end user and all their credentials. This action cannot be undone.
      parameters:
        - name: endUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: End user deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "404":
          description: End user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /end-users/{endUserId}/portal-token:
    post:
      summary: Generate portal link
      operationId: generatePortalLink
      tags:
        - End Users (Enterprise)
      description: |
        **Enterprise feature.**
        
        Generate a portal authentication link for an end user. The token expires after 30 days.
      parameters:
        - name: endUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Portal link generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      portalUrl:
                        type: string
                        format: uri
                        example: https://app.superglue.cloud/portal?token=xxx
                      expiresAt:
                        type: string
                        format: date-time
        "404":
          description: End user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /end-users/{endUserId}/invite:
    post:
      summary: Send portal invitation email
      operationId: sendPortalInvitation
      tags:
        - End Users (Enterprise)
      description: |
        **Enterprise feature.**
        
        Send a portal invitation email to the end user with a 30-day authentication link. The end user must have an email address configured.
      parameters:
        - name: endUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  recipient:
                    type: string
                    format: email
        "400":
          description: End user has no email address or email service not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: End user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: Static API Key
      description: |
        Static API key authentication using Bearer token scheme.
        Include your API key in the Authorization header: `Authorization: Bearer YOUR_API_KEY`

        Alternatively, you can use the `token` query parameter to authenticate.

        API keys can be generated in your superglue dashboard.
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - message
          properties:
            message:
              type: string
              example: Invalid input parameters
    EndUser:
      type: object
      description: |
        An end user who can connect their own credentials to multi-tenancy systems.
      required:
        - id
        - externalId
      properties:
        id:
          type: string
          format: uuid
          description: Internal superglue user ID
          example: 550e8400-e29b-41d4-a716-446655440000
        externalId:
          type: string
          description: Your application's user ID
          example: user-123
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          example: John Doe
        metadata:
          type: object
          description: Custom metadata for this user
          additionalProperties: true
        allowedSystems:
          type: array
          description: System IDs this user can access, or ["*"] for all systems
          items:
            type: string
          example: ["*"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time